AlpsNMR: Automated spectraL Processing System for NMR

AlpsNMR allows you to import NMR spectra into R and provides automated and efficient signal processing for untargeted NMR metabolomics.

Details

The following functions can be combined with the pipe. They create or modify the nmr_dataset object.

nmr_read_samples_dir() or nmr_read_samples()

nmr_interpolate_1D()

nmr_exclude_region()

nmr_normalize()

plot()

There are also functions to extract the metadata and submit the samples to irods, see the example below.

The nmr_dataset object is essentially a list, so it is easy to access its components for further analysis.

See also

Author

Maintainer: Sergio Oller Moreno sergioller@gmail.com (ORCID)

Authors:

Sergio Oller Moreno sergioller@gmail.com (ORCID)

Ivan Montoliu Roura Ivan.MontoliuRoura@rd.nestle.com

Francisco Madrid Gambin fmadrid@ibecbarcelona.eu (ORCID)

Luis Fernandez lfernandez@ibecbarcelona.eu (ORCID)

H<U+00E9>ctor Gracia Cabrera hgracia@ibecbarcelona.eu

Santiago Marco Col<U+00E1>s smarco@ibecbarcelona.eu (ORCID)

Other contributors:

Laura L<U+00F3>pez S<U+00E1>nchez [contributor]

Nestl<U+00E9> Institute of Health Sciences [copyright holder]

Institute for Bioengineering of Catalonia [copyright holder]

Miller Jack jack.miller@physics.org (ORCID) (Autophase wrapper, ASICS export) [contributor]

Examples

dir_to_demo_dataset <- system.file("dataset-demo", package = "AlpsNMR")
dataset <- nmr_read_samples_dir(dir_to_demo_dataset)
my_nmr_dataset <- dataset %>%
    nmr_interpolate_1D(axis = c(0.4, 10)) %>%
    nmr_exclude_region(exclude = list(water = c(4.6, 5))) %>%
    nmr_normalize(method = "pqn") %>%
    plot()
#> Warning: There are not enough samples for reliably estimating the median spectra
#> ℹ The Probabalistic Quotient Normalization requires several samples to compute
#>   the median spectra. Your number of samples is low
#> ℹ Review your peaks before and after normalization to ensure there are no big
#>   distortions