Deprecated function Confusion matrix of the MUVR model

After creating a model with the rdCV_PLS_RF function, you can run confusion_matrix on the model to make a confusion matrix from MUVR. This gives information about the model performance (e.g. classification rate).

confusion_matrix(MVObj, model = "mid")

Arguments

MVObj
a MUVR model

model
What type of model to plot ('min', 'mid' or 'max'). Defaults to 'mid'.

Value

A confusion matrix of the model comparing actual vs predicted class

Examples

message("Deprecated. MUVR is not compatible with Bioconductor, 
use bp_kfold_VIP_analysis method instead")

#> Deprecated. MUVR is not compatible with Bioconductor, 
#> use bp_kfold_VIP_analysis method instead