Files to rDoplhin (blood)

The rDolphin family functions are introduced to perform automatic targeted metabolite profiling. Therefore, ensure that you interpolated from -0.1 ppm in order to consider the TSP/DSS signal at 0.0 ppm. The function generates a list with the files required by to_rDolphin function. Then, it is required to save them with the save_files_to_rDolphin. to_rDolphin function will read the generated "parameters.csv" file. function.

files_to_rDolphin_blood(nmr_dataset)

Arguments

nmr_dataset
An nmr_dataset object

Value

a list containing:

meta_rDolphin: metadata in rDolphin format,

NMR_spectra: spectra matrix

ROI_blood: ROI template

Parameters_blood: parameters file

See also

Examples