NMR peak identification (cell samples)

Identify given regions and return a data frame with plausible assignations in cell samples.

nmr_identify_regions_cell(
  ppm_to_assign,
  num_proposed_compounds = 3,
  verbose = FALSE
)

Arguments

ppm_to_assign
A vector with the ppm regions to assign

num_proposed_compounds
set the number of proposed metabolites in HMDB_cell.

verbose
Logical value. Set it to TRUE to print additional information

Value

a data frame with plausible assignations.

See also

Examples

# We identify regions from from the corresponding ppm storaged in a vector.
ppm_to_assign <- c(4.060960203, 3.048970634,2.405935596,
3.24146865,0.990616851,1.002075066,0.955325548)
identification <- nmr_identify_regions_cell (ppm_to_assign, num_proposed_compounds = 3)