Search NMR samples in irods

Search for NMR samples inside irods.

nmr_irods_search(project_NPDI_ID = NA, study_nickname = NA,
  assay_ID = NA, master_sample_accession = NA,
  operational_sample_accession = NA, run_ID = NA,
  software_platform = NA, taxonomy = NA, hardware_platform = NA,
  assay_technique = NA, file_type = "Raw data file",
  file_format = "Zipped Bruker NMR directory",
  assay_platform = "Metabolomics", author = NA, exclude_trash = TRUE)

Arguments

project_NPDI_ID
The NPDI project ID (e.g. 'DUND-700713')

study_nickname
The study nickname, typically defined by a Project Manager (e.g. 'st_etienne'). A study nickname belongs to a project ID.

assay_ID
An identifier which must be unique within a study and which corresponds to a piece of work in the project management sense. It may be defined by the project manager or the platform, but should be known to the project manager. For an analytical platform it corresponds to a request to perform analysis on a set of samples from a Project. e.g. "Plasma NMR metabolites" or "Urine NMR metabolites".

master_sample_accession
The master sample accession. It is a code like 'R0000123456'.

operational_sample_accession
The operational sample accession. It is a code like 'R0000123456'.

run_ID
Optional field for operational use to separate distinct batches of samples within an Assay ID.

software_platform
The software platform used to acquire or process the data. One of NIHSrods::getValidTerms('Software platform')

taxonomy
The biological taxonomy of the sample. For instance 'Homo sapiens (Human)'. The list of defined biological taxonomies at NIHS is available here: NIHSrods::getValidTerms('Taxonomy').

hardware_platform
Machine which created the raw data. One of NIHSrods::getValidTerms('Hardware platform')

assay_technique
The analytical technique used for creating the sample. One of NIHSrods::getValidTerms('Assay technique')

file_type
By default: "Raw data file". The full list of accepted file types is available here: NIHSrods::getValidTerms('File type')

file_format
By default "Zipped Bruker NMR directory". The full list of accepted file formats is available here: NIHSrods::getValidTerms('File format')

assay_platform
By default: 'Metabolomics'. The full list of defined assay platforms is available here: NIHSrods::getValidTerms('Assay platform').

author
This will be the person who performed the processing or analysis and made the decisions. Required for processed data, the author who did the processing.

exclude_trash
Exclude samples in /NIHSData/trash (Default: TRUE)

Value

a data frame with the irods metadata that results from the search

Details

This function is slow because NIHSrods::imeta_ls does not accept multiple files at once.

See also