Create one zip file for each brucker sample path

Create one zip file for each brucker sample path

nmr_zip_bruker_samples(path, workdir, overwrite = FALSE, ...)

Arguments

path
Character vector with sample directories

workdir
Directory to store zip files

overwrite
Should existing zip files be overwritten?

...
Passed to utils::zip

Value

A character vector of the same length as path, with the zip file names

See also

Examples

outpaths <- nmr_zip_bruker_samples(".", getwd())

#> Warning: '/private/var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T/RtmpWMmGWM/filef661c471a83/reference' already exists

#> Warning: File /private/var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T/RtmpWMmGWM/filef661c471a83/reference/reference.zip already exists. Skipping