Pipeline: Add Metadata

Pipeline: Add Metadata

pipe_add_metadata(nmr_dataset_rds, excel_file, output_dir)

Arguments

nmr_dataset_rds
The nmr_dataset.rds file name coming from previous nodes

excel_file
An excel file name. See details for the requirements

The excel file can have one or more sheets. The excel sheets need to be as simple as possible: One header column on the first row and values below.

Each of the sheets contain metadata that has to be integrated. The merge (technically a left join) is done using the first column of each sheet as key.

In practical terms this means that the first sheet of the excel file MUST start with an "NMRExperiment" column, and as many additional columns to add (e.g. FluidXBarcode, SampleCollectionDate, TimePoint and SubjectID).

The second sheet can have as the first column any of the already added columns, for instance the "SubjectID", and any additional columns (e.g. Gender, Age).

The first column on each sheet, named the key column, MUST have unique values. For instance, a sheet starting with "SubjectID" MUST specify each subject ID only once (without repetitions).

output_dir
The output directory for this pipe element

Value

This function saves the result to the output directory

See also

Examples

dataset <- system.file("dataset-demo", package = "AlpsNMR")
excel_file <- system.file("dataset-demo", "dummy_metadata.xlsx", package = "AlpsNMR")
nmr_dataset <- nmr_read_samples_dir(dataset)
nmr_dataset_rds <- tempfile(fileext = ".rds")
nmr_dataset_save(nmr_dataset, nmr_dataset_rds)

#> An nmr_dataset (4 samples) 

output_dir <- tempdir()
pipe_add_metadata(nmr_dataset_rds = nmr_dataset_rds, output_dir = output_dir,
                  excel_file = excel_file)

#> Starting pipe_add_metadata at 2019-12-18 12:58:47

#> 
#> 
#> processing file: add-metadata.Rmd

#>   |                                                                         |                                                                 |   0%  |                                                                         |....                                                             |   7%
#>    inline R code fragments
#> 
#>   |                                                                         |.........                                                        |  13%
#> label: add-metadata-setup (with options) 
#> List of 1
#>  $ include: logi FALSE
#> 
#>   |                                                                         |.............                                                    |  20%
#>   ordinary text without R code
#> 
#>   |                                                                         |.................                                                |  27%
#> label: add-metadata-libraries
#>   |                                                                         |......................                                           |  33%
#>   ordinary text without R code
#> 
#>   |                                                                         |..........................                                       |  40%
#> label: add-metadata-default-values
#>   |                                                                         |..............................                                   |  47%
#>   ordinary text without R code
#> 
#>   |                                                                         |...................................                              |  53%
#> label: add-metadata-read-input
#>   |                                                                         |.......................................                          |  60%
#>   ordinary text without R code
#> 
#>   |                                                                         |...........................................                      |  67%
#> label: add-metadata-print-input
#>   |                                                                         |................................................                 |  73%
#>   ordinary text without R code
#> 
#>   |                                                                         |....................................................             |  80%
#> label: add-metadata-integrate-metadata
#>   |                                                                         |........................................................         |  87%
#>   ordinary text without R code
#> 
#>   |                                                                         |.............................................................    |  93%
#> label: add-metadata-print-output
#>   |                                                                         |.................................................................| 100%
#>   ordinary text without R code
#> 
#> 

#> output file: add-metadata.knit.md

#> "C:/Users/fmadrid/Documents/RStudio/bin/pandoc/pandoc" +RTS -K512m -RTS add-metadata.utf8.md --to html4 --from markdown+autolink_bare_uris+tex_math_single_backslash+smart --output pandoc2c1c41641bac.html --email-obfuscation none --self-contained --standalone --section-divs --template "C:\Users\fmadrid\Documents\R\R-3.6.1\library\rmarkdown\rmd\h\default.html" --no-highlight --variable highlightjs=1 --variable "theme:bootstrap" --include-in-header "C:\Users\fmadrid\AppData\Local\Temp\Rtmp8y2tJr\rmarkdown-str2c1c4f1521f3.html" --mathjax --variable "mathjax-url:https://mathjax.rstudio.com/latest/MathJax.js?config=TeX-AMS-MML_HTMLorMML" --lua-filter "C:/Users/fmadrid/Documents/R/R-3.6.1/library/rmarkdown/rmd/lua/pagebreak.lua" --lua-filter "C:/Users/fmadrid/Documents/R/R-3.6.1/library/rmarkdown/rmd/lua/latex-div.lua" 

#> 
#> Output created: C:/Users/fmadrid/AppData/Local/Temp/Rtmp8y2tJr/add-metadata.html

#> Ending pipe_add_metadata at 2019-12-18 12:58:49

# Check out: output_dir