Integrate peak positions

The function allows the integration of a given ppm vector with a specific width.

nmr_integrate_peak_positions(
  samples,
  peak_pos_ppm,
  peak_width_ppm = 0.006,
  ...
)

Arguments

samples

A nmr_dataset object

peak_pos_ppm

The peak positions, in ppm

peak_width_ppm

The peak widths (or a single peak width for all peaks)

...

Arguments passed on to nmr_integrate_regions

regions

A named list. Each element of the list is a region, given as a named numeric vector of length two with the range to integrate. The name of the region will be the name of the column

Value

Integrate peak positions

See also