Add metadata to an nmr_dataset object

This is useful to add metadata to datasets that can be later used for plotting spectra or further analysis (PCA...).

nmr_meta_add(nmr_data, metadata, by = "NMRExperiment")

nmr_meta_add_tidy_excel(nmr_data, excel_file)

Arguments

nmr_data
an nmr_dataset_family object

metadata
A data frame with metadata to add

by
A column name of both the nmr_data$metadata$external and the metadata data.frame. If you want to merge two columns with different headers you can use a named character vector c("NMRExperiment" = "ExperimentNMR") where the left side is the column name of the nmr_data$metadata$external and the right side is the column name of the metadata data frame.

excel_file
Path to a tidy Excel file name. The Excel can consist of multiple sheets, that are added sequentially. The first column of the first sheet MUST be named as one of the metadata already present in the dataset, typically will be "NMRExperiment". The rest of the columns of the first sheet can be named at will. Similary, the first column of the second sheet must be named as one of the metadata already present in the dataset, typically "NMRExperiment" or any of the columns of the first sheet. The rest of the columns of the second sheet can be named at will. See the package vignette for an example.

Value

The nmr_dataset_family object with the added metadata

See also

Examples

# Load a demo dataset with four samples:
dataset <- system.file("dataset-demo", package = "AlpsNMR")
nmr_dataset <- nmr_read_samples_dir(dataset)

#> Warning: unnest() has a new interface. See ?unnest for details.
#> Try `df %>% unnest(c(NMRExperiment, file_format, sample_path, dimension, 
#>     pulse_sequence, nuclei, acq_datetime))`, with `mutate()` if needed

#> Warning: unnest() has a new interface. See ?unnest for details.
#> Try `df %>% unnest(c(TITLE, JCAMPDX, DATATYPE, NPOINTS, ORIGIN, OWNER, ACQT0, 
#>     ANAVPT, AQSEQ, AQ_mod, AUNM, AUTOPOS, BF1, BF2, BF3, BF4, 
#>     BF5, BF6, BF7, BF8, BYTORDA, CHEMSTR, CPDPRG, DATE, DE, DECBNUC, 
#>     DECIM, DECNUC, DECSTAT, DIGMOD, DIGTYP, DQDMODE, DR, DS, 
#>     DSPFIRM, DSPFVS, DTYPA, EXP, FL1, FL2, FL3, FL4, FOV, FQ1LIST, 
#>     FQ2LIST, FQ3LIST, FQ4LIST, FQ5LIST, FQ6LIST, FQ7LIST, FQ8LIST, 
#>     FRQLO3, FRQLO3N, FTLPGN, FW, FnILOOP, FnMODE, FnTYPE, GPNAM, 
#>     GRDPROG, GRPDLY, HDDUTY, HDRATE, HL1, HL2, HL3, HL4, HOLDER, 
#>     HPPRGN, INSTRUM, LFILTER, LGAIN, LINPSTP, LOCKED, LOCKFLD, 
#>     LOCKGN, LOCKPOW, LOCKPPM, LOCNUC, LOCPHAS, LOCSHFT, LOCSW, 
#>     LTIME, MASR, MASRLST, NBL, NC, NLOGCH, NOVFLW, NS, NUC1, 
#>     NUC2, NUC3, NUC4, NUC5, NUC6, NUC7, NUC8, NUCLEUS, NUSLIST, 
#>     NusAMOUNT, NusFPNZ, NusJSP, NusSEED, NusSPTYPE, NusT2, NusTD, 
#>     O1, O2, O3, O4, O5, O6, O7, O8, OVERFLW, PAPS, PARMODE, PHLIST, 
#>     PHP, PH_ref, PLSTEP, PLSTRT, PQPHASE, PQSCALE, PR, PRGAIN, 
#>     PROBHD, PULPROG, PW, PYNM, ProjAngle, QNP, RD, RECPH, RG, 
#>     RO, SFO1, SFO2, SFO3, SFO4, SFO5, SFO6, SFO7, SFO8, SOLVENT, 
#>     SOLVOLD, SPECTR, SPINCNT, SPNAM, SUBNAM, SW, SW_h, SWfinal, 
#>     SigLockShift, TD, TD0, TDav, TE, TE1, TE2, TE3, TE4, TEG, 
#>     TE_MAGNET, TE_PIDX, TUBE_TYPE, USERA1, USERA2, USERA3, USERA4, 
#>     USERA5, V9, VALIDCODE, VALIST, VCLIST, VDLIST, VPLIST, VTLIST, 
#>     WBST, WBSW, XL, YL, YMAX_a, YMIN_a, ZGOPTNS, ZL1, ZL2, ZL3, 
#>     ZL4))`, with `mutate()` if needed

#> Warning: unnest() has a new interface. See ?unnest for details.
#> Try `df %>% unnest(c(TITLE, JCAMPDX, DATATYPE, NPOINTS, ORIGIN, OWNER, ABSF1, 
#>     ABSF2, ABSG, ABSL, ALPHA, AQORDER, ASSFAC, ASSFACI, ASSFACX, 
#>     ASSWID, AUNMP, AXLEFT, AXNAME, AXNUC, AXRIGHT, AXTYPE, AXUNIT, 
#>     AZFE, AZFW, BCFW, BC_mod, BYTORDP, COROFFS, CY, DATMOD, DC, 
#>     DFILT, DTYPP, ERETIC, F1P, F2P, FCOR, FTSIZE, FT_mod, GAMMA, 
#>     GB, INTBC, INTSCL, ISEN, LB, LEV0, LPBIN, MAXI, MC2, MEAN, 
#>     ME_mod, MI, MddCEXP, MddCT_SP, MddF180, MddLAMBDA, MddMEMORY, 
#>     MddMERGE, MddNCOMP, MddNITER, MddNOISE, MddPHASE, MddSEED, 
#>     MddSRSIZE, Mdd_CsALG, Mdd_CsLAMBDA, Mdd_CsNITER, Mdd_CsNORM, 
#>     Mdd_CsVE, Mdd_CsZF, Mdd_LrLAMBDA, Mdd_LrNITER, Mdd_LrRANK, 
#>     Mdd_LrZF, Mdd_mod, NCOEF, NC_proc, NLEV, NOISF1, NOISF2, 
#>     NSP, NTH_PI, NZP, OFFSET, PC, PHC0, PHC1, PH_mod, PKNL, PPARMOD, 
#>     PPDIAG, PPIPTYP, PPMPNUM, PPRESOL, PSCAL, PSIGN, PYNMP, RDF1, 
#>     RDF2, RDFWHM, RDINT, RDPOS, REVERSE, SF, SI, SIGF1, SIGF2, 
#>     SINO, SIOLD, SPECTYP, SREF_mod, SREGLST, SSB, STSI, STSR, 
#>     SW_p, SYMM, S_DEV, TDeff, TDoff, TI, TILT, TM1, TM2, TOPLEV, 
#>     USERP1, USERP2, USERP3, USERP4, USERP5, WDW, XDIM, YMAX_p, 
#>     YMIN_p))`, with `mutate()` if needed

# At first we just have the NMRExperiment column
print(nmr_meta_get(nmr_dataset, groups = "external"))

#> # A tibble: 4 x 1
#>   NMRExperiment
#>   <chr>        
#> 1 10           
#> 2 20           
#> 3 30           
#> 4 40           

# Get a table with NMRExperiment -> SubjectID
dummy_metadata <- system.file("dataset-demo", "dummy_metadata.xlsx", package = "AlpsNMR")
NMRExp_SubjID <- readxl::read_excel(dummy_metadata, sheet = 1)

print(NMRExp_SubjID)

#> # A tibble: 4 x 3
#>   NMRExperiment SubjectID TimePoint
#>   <chr>         <chr>     <chr>    
#> 1 10            Ana       baseline 
#> 2 20            Ana       3 months 
#> 3 30            Elia      baseline 
#> 4 40            Elia      3 months 

# We can link the SubjectID column of the first excel into the dataset
nmr_dataset <- nmr_meta_add(nmr_dataset, NMRExp_SubjID, by = "NMRExperiment")
print(nmr_meta_get(nmr_dataset, groups = "external"))

#> # A tibble: 4 x 3
#>   NMRExperiment SubjectID TimePoint
#>   <chr>         <chr>     <chr>    
#> 1 10            Ana       baseline 
#> 2 20            Ana       3 months 
#> 3 30            Elia      baseline 
#> 4 40            Elia      3 months 

# The second excel can use the SubjectID:
SubjID_Age <- readxl::read_excel(dummy_metadata, sheet = 2)
print(SubjID_Age)

#> # A tibble: 2 x 2
#>   SubjectID   Age
#>   <chr>     <dbl>
#> 1 Ana          29
#> 2 Elia          0

# Add the metadata by its SubjectID:
nmr_dataset <- nmr_meta_add(nmr_dataset, SubjID_Age, by = "SubjectID")
# The final loaded metadata:
print(nmr_meta_get(nmr_dataset, groups = "external"))

#> # A tibble: 4 x 4
#>   NMRExperiment SubjectID TimePoint   Age
#>   <chr>         <chr>     <chr>     <dbl>
#> 1 10            Ana       baseline     29
#> 2 20            Ana       3 months     29
#> 3 30            Elia      baseline      0
#> 4 40            Elia      3 months      0

# Read a tidy excel file:

dataset <- system.file("dataset-demo", package = "AlpsNMR")
nmr_dataset <- nmr_read_samples_dir(dataset)

#> Warning: unnest() has a new interface. See ?unnest for details.
#> Try `df %>% unnest(c(NMRExperiment, file_format, sample_path, dimension, 
#>     pulse_sequence, nuclei, acq_datetime))`, with `mutate()` if needed

#> Warning: unnest() has a new interface. See ?unnest for details.
#> Try `df %>% unnest(c(TITLE, JCAMPDX, DATATYPE, NPOINTS, ORIGIN, OWNER, ACQT0, 
#>     ANAVPT, AQSEQ, AQ_mod, AUNM, AUTOPOS, BF1, BF2, BF3, BF4, 
#>     BF5, BF6, BF7, BF8, BYTORDA, CHEMSTR, CPDPRG, DATE, DE, DECBNUC, 
#>     DECIM, DECNUC, DECSTAT, DIGMOD, DIGTYP, DQDMODE, DR, DS, 
#>     DSPFIRM, DSPFVS, DTYPA, EXP, FL1, FL2, FL3, FL4, FOV, FQ1LIST, 
#>     FQ2LIST, FQ3LIST, FQ4LIST, FQ5LIST, FQ6LIST, FQ7LIST, FQ8LIST, 
#>     FRQLO3, FRQLO3N, FTLPGN, FW, FnILOOP, FnMODE, FnTYPE, GPNAM, 
#>     GRDPROG, GRPDLY, HDDUTY, HDRATE, HL1, HL2, HL3, HL4, HOLDER, 
#>     HPPRGN, INSTRUM, LFILTER, LGAIN, LINPSTP, LOCKED, LOCKFLD, 
#>     LOCKGN, LOCKPOW, LOCKPPM, LOCNUC, LOCPHAS, LOCSHFT, LOCSW, 
#>     LTIME, MASR, MASRLST, NBL, NC, NLOGCH, NOVFLW, NS, NUC1, 
#>     NUC2, NUC3, NUC4, NUC5, NUC6, NUC7, NUC8, NUCLEUS, NUSLIST, 
#>     NusAMOUNT, NusFPNZ, NusJSP, NusSEED, NusSPTYPE, NusT2, NusTD, 
#>     O1, O2, O3, O4, O5, O6, O7, O8, OVERFLW, PAPS, PARMODE, PHLIST, 
#>     PHP, PH_ref, PLSTEP, PLSTRT, PQPHASE, PQSCALE, PR, PRGAIN, 
#>     PROBHD, PULPROG, PW, PYNM, ProjAngle, QNP, RD, RECPH, RG, 
#>     RO, SFO1, SFO2, SFO3, SFO4, SFO5, SFO6, SFO7, SFO8, SOLVENT, 
#>     SOLVOLD, SPECTR, SPINCNT, SPNAM, SUBNAM, SW, SW_h, SWfinal, 
#>     SigLockShift, TD, TD0, TDav, TE, TE1, TE2, TE3, TE4, TEG, 
#>     TE_MAGNET, TE_PIDX, TUBE_TYPE, USERA1, USERA2, USERA3, USERA4, 
#>     USERA5, V9, VALIDCODE, VALIST, VCLIST, VDLIST, VPLIST, VTLIST, 
#>     WBST, WBSW, XL, YL, YMAX_a, YMIN_a, ZGOPTNS, ZL1, ZL2, ZL3, 
#>     ZL4))`, with `mutate()` if needed

#> Warning: unnest() has a new interface. See ?unnest for details.
#> Try `df %>% unnest(c(TITLE, JCAMPDX, DATATYPE, NPOINTS, ORIGIN, OWNER, ABSF1, 
#>     ABSF2, ABSG, ABSL, ALPHA, AQORDER, ASSFAC, ASSFACI, ASSFACX, 
#>     ASSWID, AUNMP, AXLEFT, AXNAME, AXNUC, AXRIGHT, AXTYPE, AXUNIT, 
#>     AZFE, AZFW, BCFW, BC_mod, BYTORDP, COROFFS, CY, DATMOD, DC, 
#>     DFILT, DTYPP, ERETIC, F1P, F2P, FCOR, FTSIZE, FT_mod, GAMMA, 
#>     GB, INTBC, INTSCL, ISEN, LB, LEV0, LPBIN, MAXI, MC2, MEAN, 
#>     ME_mod, MI, MddCEXP, MddCT_SP, MddF180, MddLAMBDA, MddMEMORY, 
#>     MddMERGE, MddNCOMP, MddNITER, MddNOISE, MddPHASE, MddSEED, 
#>     MddSRSIZE, Mdd_CsALG, Mdd_CsLAMBDA, Mdd_CsNITER, Mdd_CsNORM, 
#>     Mdd_CsVE, Mdd_CsZF, Mdd_LrLAMBDA, Mdd_LrNITER, Mdd_LrRANK, 
#>     Mdd_LrZF, Mdd_mod, NCOEF, NC_proc, NLEV, NOISF1, NOISF2, 
#>     NSP, NTH_PI, NZP, OFFSET, PC, PHC0, PHC1, PH_mod, PKNL, PPARMOD, 
#>     PPDIAG, PPIPTYP, PPMPNUM, PPRESOL, PSCAL, PSIGN, PYNMP, RDF1, 
#>     RDF2, RDFWHM, RDINT, RDPOS, REVERSE, SF, SI, SIGF1, SIGF2, 
#>     SINO, SIOLD, SPECTYP, SREF_mod, SREGLST, SSB, STSI, STSR, 
#>     SW_p, SYMM, S_DEV, TDeff, TDoff, TI, TILT, TM1, TM2, TOPLEV, 
#>     USERP1, USERP2, USERP3, USERP4, USERP5, WDW, XDIM, YMAX_p, 
#>     YMIN_p))`, with `mutate()` if needed

# At first we just have the NMRExperiment column
print(nmr_meta_get(nmr_dataset, groups = "external"))

#> # A tibble: 4 x 1
#>   NMRExperiment
#>   <chr>        
#> 1 10           
#> 2 20           
#> 3 30           
#> 4 40           

# Get a table with NMRExperiment -> SubjectID
dummy_metadata <- system.file("dataset-demo", "dummy_metadata.xlsx", package = "AlpsNMR")

nmr_dataset <-nmr_meta_add_tidy_excel(nmr_dataset, dummy_metadata)
# Updated Metadata:
print(nmr_meta_get(nmr_dataset, groups = "external"))

#> # A tibble: 4 x 4
#>   NMRExperiment SubjectID TimePoint   Age
#>   <chr>         <chr>     <chr>     <dbl>
#> 1 10            Ana       baseline     29
#> 2 20            Ana       3 months     29
#> 3 30            Elia      baseline      0
#> 4 40            Elia      3 months      0