Random subsampling

Random subsampling

random_subsampling(
  sample_idx,
  iterations = 10L,
  test_size = 0.25,
  keep_together = NULL,
  balance_in_train = NULL
)

Arguments

sample_idx

Typically a numeric vector with sample index to be separated. A character vector with sample IDs could also be used

iterations

An integer, the number of iterations in the random subsampling

test_size

A number between 0 and 1. The samples to be included in the test set on each interation.

keep_together

Either NULL or a factor with the same length as sample_idx. keep_together can be used to ensure that groups of samples are kept in together in all iterations (either on training or on test, but never split). A typical use case for this is when you have sample replicates and you want to keep all replicates together to prevent overoptimistic results (having one sample on the train subset and its replicate on the test subset would make the prediction easier to guess). Another use case for this is when you have a longitudinal study and you want to keep some subjects in the same train or test group, because you want to use some information in a longitudinal way (e.g. a multilevel plsda model).

balance_in_train

Either NULL or a factor with the same length as sample_idx. balance_in_train can be used to force that on each iteration, the train partition contains the same number of samples of the given factor levels. For instance, if we have a dataset with 40 samples of class "A" and 20 samples of class "B", using a test_size = 0.25, we can force to always have 16 samples of class "A" and 16 samples of class "B" in the training subset. This is beneficial to those algorithms that require that the training groups are balanced.

Value

A list of length equal to iterations. Each element of the list is a list with two entries (training and test) containing the sample_idxvalues that will belong to each subset.

Examples

random_subsampling(1:100, iterations = 4, test_size = 0.25)
#> [[1]]
#> [[1]]$training
#>  [1]   3   4   5   7   8   9  10  11  12  13  15  16  18  19  20  22  23  24  25
#> [20]  26  27  28  29  30  32  33  34  35  36  37  38  40  41  42  43  44  45  47
#> [39]  48  49  50  51  52  53  54  55  56  57  58  59  60  62  64  66  68  69  70
#> [58]  71  72  75  77  78  79  82  83  87  88  89  92  93  95  97  98  99 100
#> 
#> [[1]]$test
#>  [1]  1  2  6 14 17 21 31 39 46 61 63 65 67 73 74 76 80 81 84 85 86 90 91 94 96
#> 
#> 
#> [[2]]
#> [[2]]$training
#>  [1]   1   2   4   5   6   8   9  10  11  12  13  15  16  17  18  19  20  21  22
#> [20]  23  24  25  27  28  29  30  32  33  34  35  36  37  39  40  41  42  43  45
#> [39]  47  49  50  51  52  53  55  56  57  58  62  63  64  65  66  67  69  70  73
#> [58]  74  75  76  78  79  80  81  82  83  85  86  89  90  91  93  96  98 100
#> 
#> [[2]]$test
#>  [1]  3  7 14 26 31 38 44 46 48 54 59 60 61 68 71 72 77 84 87 88 92 94 95 97 99
#> 
#> 
#> [[3]]
#> [[3]]$training
#>  [1]   2   3   4   5   6   8  10  13  14  15  16  17  18  19  21  22  23  25  26
#> [20]  27  28  29  33  35  36  38  40  41  43  44  45  46  47  50  51  52  54  55
#> [39]  56  57  58  59  60  61  62  63  64  65  66  68  69  70  72  74  75  76  78
#> [58]  79  80  81  82  83  84  85  86  87  89  90  91  92  94  95  96  98 100
#> 
#> [[3]]$test
#>  [1]  1  7  9 11 12 20 24 30 31 32 34 37 39 42 48 49 53 67 71 73 77 88 93 97 99
#> 
#> 
#> [[4]]
#> [[4]]$training
#>  [1]  1  2  3  4  5  6  7  8  9 10 11 12 13 14 15 16 17 18 19 20 21 22 23 24 25
#> [26] 26 27 28 29 30 31 32 33 34 36 39 40 43 45 46 48 50 51 52 53 54 56 59 61 62
#> [51] 63 69 70 71 72 73 74 76 78 79 82 83 84 85 87 88 89 90 91 92 93 95 96 97 99
#> 
#> [[4]]$test
#>  [1]  35  37  38  41  42  44  47  49  55  57  58  60  64  65  66  67  68  75  77
#> [20]  80  81  86  94  98 100
#> 
#> 

subject_id <- c("Alice", "Bob", "Charlie", "Eve")
random_subsampling(1:4, iterations = 2, test_size = 0.25, keep_together = subject_id)
#> [[1]]
#> [[1]]$training
#> [1] 1 3 4
#> 
#> [[1]]$test
#> [1] 2
#> 
#> 
#> [[2]]
#> [[2]]$training
#> [1] 1 2 3
#> 
#> [[2]]$test
#> [1] 4
#> 
#>